Fetch interaction initial values via task
Gets initial values through task mechanism rather than in controller
