+++ src/apps/interactions/apps/details-form/client/state.js
@@ -3,3 +3,5 @@ export const ID = 'interactionDetailsForm'
 export const TASK_SAVE_INTERACTION = 'TASK_SAVE_INTERACTION'
 export const TASK_OPEN_CONTACT_FORM = 'TASK_OPEN_CONTACT_FORM'
 export const TASK_GET_ACTIVE_EVENTS = 'TASK_GET_ACTIVE_EVENTS'
+export const TASK_GET_INTERACTION_INITIAL_VALUES =
+  'TASK_GET_INTERACTION_INITIAL_VALUES'
