+++ src/client/provider.jsx
@@ -43,7 +43,6 @@ import { ID as EXPORT_COUNTRIES_EDIT_ID } from '../apps/companies/apps/exports/c
 import exportCountriesEditReducer from '../apps/companies/apps/exports/client/ExportCountriesEdit/reducer'
 
 import * as addInteractionFormState from '../apps/interactions/apps/details-form/client/state'
-import * as addInteractionFormTasks from '../apps/interactions/apps/details-form/client/tasks'
 import addInteractionFormReducer from '../apps/interactions/apps/details-form/client/reducer'
 
 import * as addCompanyState from '../apps/companies/apps/add-company/client/state'
@@ -201,15 +200,6 @@ const store = createStore(
   {
     referrerUrl: window.document.referrer,
     Form: {
-      [addInteractionFormState.ID]: {
-        values: {},
-        touched: {},
-        errors: {},
-        fields: {},
-        steps: [],
-        currentStep: 0,
-        ...addInteractionFormTasks.restoreState(),
-      },
       [REFERRALS_SEND_ID]: {
         values: {},
         touched: {},
